requirements.txt
asgiref==3.5.2
Django==4.1.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
